pyproject.toml
[tool.poetry]
name = "db-atbasica"
version = "0.1.0"
description = ""
authors = ["gleyce-alves <gleyceaps.alves@gmail.com>", "plz09 <r.pellizzi09@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
sqlalchemy = "^2.0.31"
psycopg2 = "^2.9.9"
unidecode = "^1.3.8"
openpyxl = "^3.1.5"
duckdb = "^1.0.0"
office365-rest-python-client = "^2.5.11"
requests = "^2.32.3"
requests-ntlm = "^1.3.0"
python-dotenv = "^1.0.1"
pyarrow = "^17.0.0"
dask = "^2024.7.1"
pyspark = "^3.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
